+++ CesiumGltf/src/Accessor.cpp
@@ -4,6 +4,7 @@
 
 #include <CesiumGltf/BufferView.h>
 
+#include <cstddef>
 #include <cstdint>
 #include <string>
 
@@ -82,7 +83,8 @@ Accessor::computeByteStride(const CesiumGltf::Model& model) const noexcept {
   if (pBufferView->byteStride && pBufferView->byteStride.value() != 0) {
     return pBufferView->byteStride.value();
   }
-  return computeNumberOfComponents(this->type) *
-         computeByteSizeOfComponent(this->componentType);
+  return static_cast<int64_t>(
+      computeNumberOfComponents(this->type) *
+      computeByteSizeOfComponent(this->componentType));
 }
 } // namespace CesiumGltf
