+++ CesiumRasterOverlays/include/CesiumRasterOverlays/RasterOverlay.h
@@ -14,6 +14,7 @@
 #include <spdlog/fwd.h>
 
 #include <any>
+#include <cstddef>
 #include <functional>
 #include <memory>
 #include <optional>
@@ -50,7 +51,7 @@ struct CESIUMRASTEROVERLAYS_API RasterOverlayOptions {
    * in memory in case they're needed again soon. This property controls the
    * maximum size of that cache.
    */
-  int64_t subTileCacheBytes = 16 * 1024 * 1024;
+  int64_t subTileCacheBytes = static_cast<int64_t>(16 * 1024 * 1024);
 
   /**
    * @brief The maximum pixel size of raster overlay textures, in either
